+++ src/tech/v3/datatype_api.clj
@@ -24,7 +24,9 @@
             [tech.v3.datatype.io-indexed-buffer]
             [tech.v3.datatype.const-reader]
             [tech.v3.datatype.io-concat-buffer]
-            [tech.v3.datatype.copy-make-container])
+            [tech.v3.datatype.copy-make-container]
+            [tech.v3.datatype.fastobjs :as fastobjs]
+            [tech.v3.datatype.imlist :as imlist])
   (:import [tech.v3.datatype ListPersistentVector BooleanReader
             LongReader DoubleReader ObjectReader PrimitiveList
             Buffer ArrayBufferData NativeBufferData]
@@ -323,6 +325,33 @@ user> (dtype/make-reader :float32 5 (* idx 2))
                 retval
                 (errors/throwf "Unable create create tensor for nd object type: %s"
                                (type item))))))
+
+
+(defn map-factory
+  "Create an IFn taking exactly n-keys arguments to rapidly create a map.
+  This moves the key-checking to the factory creation and simply creates a
+  new map as fast as possible when requrested"
+  [key-seq]
+  (if (empty? key-seq)
+    (constantly {})
+    (fastobjs/map-factory key-seq)))
+
+
+(defn fast-vector
+  "Much faster vector/list constructor than `clojure.core/vector`."
+  ([] [])
+  ([a0] (imlist/imlist a0))
+  ([a0 a1] (imlist/imlist a0 a1))
+  ([a0 a1 a2] (imlist/imlist a0 a1 a2))
+  ([a0 a1 a2 a3] (imlist/imlist a0 a1 a2 a3))
+  ([a0 a1 a2 a3 a4] (imlist/imlist a0 a1 a2 a3 a4))
+  ([a0 a1 a2 a3 a4 a5] (imlist/imlist a0 a1 a2 a3 a4 a5))
+  ([a0 a1 a2 a3 a4 a5 a6] (imlist/imlist a0 a1 a2 a3 a4 a5 a6))
+  ([a0 a1 a2 a3 a4 a5 a6 a7] (imlist/imlist a0 a1 a2 a3 a4 a5 a6 a7))
+  ([a0 a1 a2 a3 a4 a5 a6 a7 a8] (imlist/imlist a0 a1 a2 a3 a4 a5 a6 a7 a8))
+  ([a0 a1 a2 a3 a4 a5 a6 a7 a8 a9] (imlist/imlist a0 a1 a2 a3 a4 a5 a6 a7 a8 a9))
+  ([a0 a1 a2 a3 a4 a5 a6 a7 a8 a9 & args]
+   (apply imlist/imlist a0 a1 a2 a3 a4 a5 a6 a7 a8 a9 args)))
 
 (comment